requirements.txt
# GraphRAG系统依赖包
neo4j==5.15.0
requests==2.31.0
flask==3.0.0
python-dotenv==1.0.0
typing-extensions==4.8.0
# dataclasses是Python 3.7+内置模块，无需单独安装

# 可选依赖（用于更好的性能和功能）
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2

# 开发和测试依赖
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Web界面相关
jinja2==3.1.2
werkzeug==3.0.1